+++ lib/Sema/ConstraintSystem.h
@@ -18,7 +18,6 @@
 #ifndef SWIFT_SEMA_CONSTRAINT_SYSTEM_H
 #define SWIFT_SEMA_CONSTRAINT_SYSTEM_H
 
-#include "CSFix.h"
 #include "Constraint.h"
 #include "swift/AST/ASTContext.h"
 #include "swift/AST/ASTNode.h"
@@ -35,6 +34,7 @@
 #include "swift/Sema/ConstraintGraph.h"
 #include "swift/Sema/ConstraintGraphScope.h"
 #include "swift/Sema/ConstraintLocator.h"
+#include "swift/Sema/CSFix.h"
 #include "swift/Sema/OverloadChoice.h"
 #include "swift/Sema/SolutionResult.h"
 #include "llvm/ADT/MapVector.h"
